MSALinearity-class: Class '"MSALinearity"'

R6 class for performing Measurement System Analysis (MSA) Linearity studies.

Public fields

X

A data frame containing the independent variable(s) used in the linearity study.

Y

A data frame containing the dependent variable(s) or responses measured in the linearity study.

model

The linear model object resulting from the linearity analysis.

conf.level

A numeric value specifying the confidence level for the linearity analysis. This should be between 0 and 1 (e.g., 0.95 for a 95% confidence level).

Linearity

A list or data frame containing the results of the linearity study, including the linearity value and associated statistics.

GageName

A character string specifying the name of the gage or measurement system under analysis.

GageTolerance

A numeric value specifying the tolerance of the gage or measurement system.

DateOfStudy

A character string or Date object indicating the date when the linearity study was conducted.

PersonResponsible

A character string specifying the name of the person responsible for the linearity study.

Comments

A character string for additional comments or notes about the linearity study.

facNames

A character vector specifying the names of the factors involved in the study, if any.

Methods

Public methods

MSALinearity$response()

MSALinearity$summary()

MSALinearity$plot()

MSALinearity$print()

MSALinearity$as.data.frame()

MSALinearity$clone()

Method response()

Get and set the the response in an object of class MSALinearity.

Usage

MSALinearity$response(value)

Arguments

value

New response, If missing value get the response.

Method summary()

Methods for function summary in Package base.

Usage

MSALinearity$summary()

Method plot()

Plots the measurement system, including individual biases, mean bias, and a regression line with confidence intervals.

Usage

MSALinearity$plot(ylim, col, pch, lty = c(1, 2), ...)

Arguments

ylim

A numeric vector specifying the limits for the y-axis. If not provided, the limits are automatically calculated based on data.

col

A vector specifying the colors to be used for different plot elements.

pch

A numeric vector specifying the plotting characters (symbols) for individual data points and mean bias points.

lty

A numeric vector specifying the line types for the regression line and its confidence intervals. The default is c(1, 2).

...

Additional graphical parameters passed to the plot function, allowing further customization of the plot's appearance.

Method print()

Methods for function print in Package base.

Usage

MSALinearity$print()

Method as.data.frame()

Return a data frame with the information of the object MSALinearity.

Usage

MSALinearity$as.data.frame()

Method clone()

The objects of this class are cloneable with this method.

Usage

MSALinearity$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.