confounds: Confounded Effects

Function to display confounded effects of a fractional factorial design in a human readable way.

Usage

confounds(x, depth = 2)

Arguments

x

An object of class facDesign.c.

depth

numeric value - up to depth-way confounded interactions are printed

Value

The function returns a summary of the factors confounded.

Examples

vp.frac = fracDesign(k = 4, gen = "D=ABC")
confounds(vp.frac,depth=5)
#> Defining relations:
#> I =  ABCD 		Columns: 1 2 3 4 
#> 
#> Resolution:  IV 
#> 
#> Alias Structure:
#> A 	is confounded with	 BCD 
#> B 	is confounded with	 ACD 
#> C 	is confounded with	 ABD 
#> D 	is confounded with	 ABC 
#> AB 	is confounded with	 CD 
#> AC 	is confounded with	 BD 
#> AD 	is confounded with	 BC 
#> ABCD 	is confounded with