desOpt-class: Class `desOpt`

The desOpt class represents an object that stores optimization results for factorial design experiments. It includes coded and real factors, responses, desirabilities, overall desirability, and the design object.

See also

Public fields

facCoded

A list containing the coded values for the factors in the design.

facReal

A list containing the real (actual) values for the factors in the design.

responses

A list of response variables obtained from the design.

desirabilities

A list of desirability scores for each response variable.

overall

Numeric value representing the overall desirability score.

all

A data frame containing all the relevant data from the design and optimization process.

fdo

The factorial design object used in the optimization process.

Methods

Public methods

desOpt$as.data.frame()

desOpt$print()

desOpt$clone()

Method as.data.frame()

Convert the object to a data frame.

Usage

desOpt$as.data.frame()

Method print()

Print a summary of the object.

Usage

desOpt$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

desOpt$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.