desirability-class: Class "desirability"

A class representing the desirability metrics for responses in a design.

See also

Public fields

response

A numeric vector specifying the responses for which desirability is calculated.

low

A numeric vector representing the lower bounds of the desirable range for each response.

high

A numeric vector representing the upper bounds of the desirable range for each response.

target

A numeric vector or character string indicating the target values or goals for each response.

scale

A numeric vector specifying the scaling factors used in the desirability calculation.

importance

A numeric vector indicating the importance of each response in the desirability calculation.

Methods

Public methods

desirability.c$new()

desirability.c$print()

desirability.c$plot()

desirability.c$clone()

Method new()

Initializes a new desirability.c object with specified parameters.

Usage

desirability.c$new(
  response = NULL,
  low = NULL,
  high = NULL,
  target = NULL,
  scale = NULL,
  importance = NULL
)

Arguments

response

A numeric or character vector specifying the responses for which desirability is calculated.

low

A numeric vector representing the lower bounds of the desirable range for each response.

high

A numeric vector representing the upper bounds of the desirable range for each response.

target

A numeric vector or character string indicating the target values or goals for each response.

scale

A numeric vector specifying the scaling factors used in the desirability calculation.

importance

A numeric vector indicating the importance of each response in the desirability calculation.

Method print()

Prints the details of a desirability.c object.

Usage

desirability.c$print()

Method plot()

Plots the desirability functions based on the specified parameters.

Usage

desirability.c$plot(scale, main, xlab, ylab, line.width, col, numPoints = 500)

Arguments

scale

A numeric vector specifying the scaling factors used in the plot.

main

A character string specifying the main title of the plot.

xlab

A character string specifying the label for the x-axis.

ylab

A character string specifying the label for the y-axis.

line.width

A numeric value specifying the width of the plot lines.

col

A vector of colors for the plot lines.

numPoints

An integer specifying the number of points to plot (default is 500).

Method clone()

The objects of this class are cloneable with this method.

Usage

desirability.c$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.