distribution: Distribution

Calculates the most likely parameters for a given distribution.

Usage

distribution(x = NULL, distrib = "weibull", ...)

Arguments

x

Vector of distributed values from which the parameter should be determined.

distrib

Character string specifying the distribution of x. The function distribution will accept the following character strings for distribution:

`normal`

`chi-squared`

`exponential`

`logistic`

`gamma`

`weibull`

`cauchy`

`beta`

`f`

`t`

`geometric`

`poisson`

`negative binomial`

`log-normal`

By default, distribution is set to `weibull`.

...

Additional arguments to be passed to the fitting function.

Value

distribution() returns an object of class DistrCollection.

See also

Examples

data1 <- rnorm(100, mean = 5, sd = 2)
distribution(data1, distrib = "normal")
#> 
#> 
#> fitted distribution is normal :
#>     mean       sd 
#> 5.030286 1.860162 
#>