dlnorm3: The Lognormal Distribution (3 Parameter)

Density function, distribution function, and quantile function for the Lognormal distribution.

dlnorm3(x, meanlog, sdlog, threshold)
plnorm3(q, meanlog, sdlog, threshold)
qlnorm3(p, meanlog, sdlog, threshold, ...)

Arguments

x, q

A numeric vector of quantiles.

meanlog, sdlog

The mean and standard deviation of the distribution on the log scale with default values of 0 and 1 respectively.

threshold

The threshold parameter, default is 0.

p

A numeric vector of probabilities.

...

Additional arguments that can be passed to uniroot.

Value

dlnorm3 gives the density, plnorm3 gives the distribution function, and qlnorm3 gives the quantile function.

Details

The Lognormal distribution with meanlog parameter zeta, sdlog parameter sigma, and threshold parameter theta has a density given by:

$$f(x) = \frac{1}{\sqrt{2\pi}\sigma(x-\theta)}\exp\left(-\frac{(\log(x-\theta)-\zeta)^2}{2\sigma^2}\right)$$

The cumulative distribution function is given by:

$$F(x) = \Phi\left(\frac{\log(x-\theta)-\zeta}{\sigma}\right)$$

where \(\Phi\) is the cumulative distribution function of the standard normal distribution.

Examples

dlnorm3(x = 2, meanlog = 0, sdlog = 1/8, threshold = 1)
#> [1] 3.191538
temp <- plnorm3(q = 2, meanlog = 0, sdlog = 1/8, threshold = 1)
temp
#> [1] 0.5
qlnorm3(p = temp, meanlog = 0, sdlog = 1/8, threshold = 1)
#> [1] 2.000004