doeFactor-class: Class `doeFactor`

An R6 class representing a factor in a design of experiments (DOE).

See also

Public fields

low

Numeric value specifying the lower bound of the factor. Default is `-1`.

high

Numeric value specifying the upper bound of the factor. Default is `1`.

name

Character string specifying the name of the factor. Default is an empty string ``.

unit

Character string specifying the unit of measurement for the factor. Default is an empty string ``.

type

Character string specifying the type of the factor. Can be either `numeric` or `categorical`. Default is `numeric`.

Methods

Public methods

doeFactor$attributes()

doeFactor$.low()

doeFactor$.high()

doeFactor$.type()

doeFactor$.unit()

doeFactor$names()

doeFactor$print()

doeFactor$clone()

Method attributes()

Get the attributes of the factor.

Usage

doeFactor$attributes()

Method .low()

Get and set the lower bound for the factor.

Usage

doeFactor$.low(value)

Arguments

value

Numeric value to set as the lower bound. If missing, the current lower bound is returned.

Method .high()

Get and set the upper bound for the factor.

Usage

doeFactor$.high(value)

Arguments

value

Numeric value to set as the upper bound. If missing, the current upper bound is returned.

Method .type()

Get and set the type of the factor.

Usage

doeFactor$.type(value)

Arguments

value

Character string specifying the type of the factor. Can be `"numeric"` or `"categorical"`. If missing, the current type is returned.

Method .unit()

Get and set the unit of measurement for the factor.

Usage

doeFactor$.unit(value)

Arguments

value

Character string specifying the unit of measurement. If missing, the current unit is returned.

Method names()

Get and set the name of the factor.

Usage

doeFactor$names(value)

Arguments

value

Character string specifying the name of the factor. If missing, the current name is returned.

Method print()

Print the characteristics of the factors.

Usage

doeFactor$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

doeFactor$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.