pbDesign: Plackett-Burman Designs

Function to create a Plackett-Burman design.

Usage

pbDesign(n, k, randomize = TRUE, replicates = 1)

Arguments

n

Integer value giving the number of trials.

k

Integer value giving the number of factors.

randomize

A logical value (TRUE/FALSE) that specifies whether to randomize the RunOrder of the design. By default, randomize is set to TRUE.

replicates

An integer specifying the number of replicates for each run in the design.

Value

A pbDesign returns an object of class pbDesign.

Note

This function creates Placket-Burman Designs down to n=26. Bigger Designs are not implemented because of lack in practicability. For the creation either the number of factors or the number of trials can be denoted. Wrong combinations will lead to an error message. Originally Placket-Burman-Design are applicable for number of trials divisible by 4. If n is not divisble by 4 this function will take the next larger Placket-Burman Design and truncate the last rows and columns.

See also

Examples

pbdo<- pbDesign(n=5)
pbdo$summary()
#> Plackett-Burman  Design
#> Information about the factors:
#> 
#>               A       B       C       D
#> value 1      -1      -1      -1      -1
#> value 2       1       1       1       1
#> name                                   
#> unit                                   
#> type    numeric numeric numeric numeric
#> 
#> -----------
#> 
#>   StandOrder RunOrder Replicate  A  B  C  D  y
#> 1          4        1         1  1  1 -1  1 NA
#> 2          5        2         1 -1 -1 -1 -1 NA
#> 3          1        3         1  1  1  1 -1 NA
#> 4          2        4         1 -1  1  1  1 NA
#> 5          3        5         1  1 -1  1  1 NA
#> 
#> -----------
#>