pbDesign

An R6 class representing a Plackett-Burman design.

Public fields

name

A character string specifying the name of the design. Default is NULL.

factors

A list of factors included in the Taguchi design. Each factor is typically an instance of the pbFactor class.

design

A data.frame representing the design matrix of the experiment. This includes the levels of each factor for every run of the experiment. Default is an empty data.frame.

designType

A character string specifying the type of Taguchi design used. Default is NULL.

replic

A data.frame containing the replication information for the design. Default is an empty data.frame.

response

A data.frame storing the response values collected from the experiment. Default is an empty data.frame.

Type

A data.frame specifying the type of responses or factors involved in the design. Default is an empty data.frame.

block

A data.frame indicating any blocking factors used in the design. Default is an empty data.frame.

runOrder

A data.frame detailing the order in which the experimental runs were conducted. Default is an empty data.frame.

standardOrder

A data.frame detailing the standard order of the experimental runs. Default is an empty data.frame.

desireVal

A list storing desired values for responses in the experiment. Default is an empty list.

desirability

A list storing desirability functions used to evaluate the outcomes of the experiment. Default is an empty list.

fits

A data.frame containing model fits or other statistical summaries from the analysis of the experimental data. Default is an empty data.frame.

Methods

Public methods

pbDesign.c$values()

pbDesign.c$units()

pbDesign.c$.factors()

pbDesign.c$names()

pbDesign.c$as.data.frame()

pbDesign.c$print()

pbDesign.c$.response()

pbDesign.c$.nfp()

pbDesign.c$summary()

pbDesign.c$clone()

Method values()

Get and set the values for an object of class pbDesign.

Usage

pbDesign.c$values(value)

Arguments

value

New value, If missing value get the values.

Method units()

Get and set the units for an object of class pbDesign.

Usage

pbDesign.c$units(value)

Arguments

value

New units, If missing value get the units.

Method .factors()

Get and set the factors in an object of class pbDesign.

Usage

pbDesign.c$.factors(value)

Arguments

value

New factors, If missing value get the factors.

Method names()

Get and set the names in an object of class pbDesign.

Usage

pbDesign.c$names(value)

Arguments

value

New names, If missing value get the names.

Method as.data.frame()

Return a data frame with the information of the object pbDesign.

Usage

pbDesign.c$as.data.frame()

Method print()

Methods for function print in Package base.

Usage

pbDesign.c$print()

Method .response()

Get and set the the response in an object of class pbDesign.

Usage

pbDesign.c$.response(value)

Arguments

value

New response, If missing value get the response.

Method .nfp()

Prints a summary of the factors attributes including their low, high, name, unit, and type.

Usage

pbDesign.c$.nfp()

Method summary()

Methods for function summary in Package base.

Usage

pbDesign.c$summary()

Method clone()

The objects of this class are cloneable with this method.

Usage

pbDesign.c$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.