pbFactor

An R6 class representing a factor in a Plackett-Burman design.

Public fields

values

A vector containing the levels or values associated with the factor. Default is NA.

name

A character string specifying the name of the factor. Default is an empty string "".

unit

A character string specifying the unit of measurement for the factor. Default is an empty string "".

type

A character string specifying the type of the factor, which can be either "numeric" or "categorical". Default is "numeric".

Methods

Public methods

pbFactor$attributes()

pbFactor$.values()

pbFactor$.unit()

pbFactor$names()

pbFactor$clone()

Method attributes()

Get the attributes of the factor.

Usage

pbFactor$attributes()

Method .values()

Get and set the values for the factors in an object of class pbFactor.

Usage

pbFactor$.values(value)

Arguments

value

New values, If missing value get the values.

Method .unit()

Get and set the units for the factors in an object of class pbFactor.

Usage

pbFactor$.unit(value)

Arguments

value

New unit, If missing value get the units.

Method names()

Get and set the names in an object of class pbFactor.

Usage

pbFactor$names(value)

Arguments

value

New names, If missing value get the names.

Method clone()

The objects of this class are cloneable with this method.

Usage

pbFactor$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.