qgamma3: The gamma Distribution (3 Parameter)

Density function, distribution function, and quantile function for the Gamma distribution.

Usage

qgamma3(p, shape, scale, threshold, ...)

Arguments

p

A numeric vector of probabilities.

shape

The shape parameter, default is 1.

scale

The scale parameter, default is 1.

threshold

The threshold parameter, default is 0.

...

Additional arguments that can be passed to uniroot.

Value

dgamma3 gives the density, pgamma3 gives the distribution function, and qgamma3 gives the quantile function.

Details

The Gamma distribution with scale parameter alpha, shape parameter c, and threshold parameter zeta has a density given by:

$$f(x) = \frac{c}{\alpha}\left(\frac{x-\zeta}{\alpha}\right)^{c-1}\exp\left(-\left(\frac{x-\zeta}{\alpha}\right)^c\right)$$

The cumulative distribution function is given by:

$$F(x) = 1 - \exp\left(-\left(\frac{x-\zeta}{\alpha}\right)^c\right)$$

Examples

dgamma3(x = 1, scale = 1, shape = 5, threshold = 0)
#> [1] 0.01532831
temp <- pgamma3(q = 1, scale = 1, shape = 5, threshold = 0)
temp
#> [1] 0.003659847
qgamma3(p = temp, scale = 1, shape = 5, threshold = 0)
#> [1] 0.9999981