simProc: Simulated Process

This is a function to simulate a black box process for teaching the use of designed experiments. The optimal factor settings can be found using a sequential assembly strategy i.e. apply a 2^k factorial design first, calculate the path of the steepest ascent, again apply a 2^k factorial design and augment a star portion to find the optimal factor settings. Of course, other strategies are possible.

Usage

simProc(x1, x2, x3, noise = TRUE)

Arguments

x1

numeric vector containing the values for factor 1.

x2

numeric vector containing the values for factor 2.

x3

numeric vector containing the values for factor 3.

noise

logical value deciding whether noise should be added or not. Default setting is TRUE.

Value

simProc returns a numeric value within the range [0,1].

Examples

simProc(120, 140, 1)
#> [1] 0.4777199
simProc(120, 220, 1)
#> [1] 0.00209724
simProc(160, 140, 1)
#> [1] 0.2638334