starDesign: Axial Design

starDesign is a function to create the star portion of a response surface design. The starDesign function can be used to create a star portion of a response surface design for a sequential assembly strategy. One can either specify k and p and alpha and cs and cc OR simply simply pass an object of class facDesign.c to the data. In the latter an object of class facDesign.c otherwise a list containing the axial runs and centerpoints is returned.

Usage

starDesign(
  k,
  p = 0,
  alpha = c("both", "rotatable", "orthogonal"),
  cs,
  cc,
  data
)

Arguments

k

Integer value giving number of factors.

p

Integer value giving the number of factors via aliasing. By default set to `0`.

alpha

If no numeric value is given defaults to `both` i.e. `orthogonality` and `rotatibility` which can be set as character strings too.

cs

Integer value giving the number of centerpoints in the star portion of the design.

cc

Integer value giving the number of centerpoints in the cube portion of the design.

data

Optional. An object of class facDesign.c.

Value

starDesign returns a facDesign.c object if an object of class facDesign.c is given or a list containing entries for axial runs and center points in the cube and the star portion of a design.

See also

Examples

# Example 1: sequential assembly
# Factorial design with one center point in the cube portion
fdo = facDesign(k = 3, centerCube = 1)
# Set the response via generic response method
fdo$.response(1:9)
# Sequential assembly of a response surface design (rsd)
rsd = starDesign(data = fdo)

# Example 2: Returning a list of star point designs
starDesign(k = 3, cc = 2, cs = 2, alpha = "orthogonal")
#> $star
#>           A         B         C
#> 1 -1.788854  0.000000  0.000000
#> 2  1.788854  0.000000  0.000000
#> 3  0.000000 -1.788854  0.000000
#> 4  0.000000  1.788854  0.000000
#> 5  0.000000  0.000000 -1.788854
#> 6  0.000000  0.000000  1.788854
#> 
#> $centerStar
#>   A B C
#> 1 0 0 0
#> 2 0 0 0
#> 
#> $centerCube
#>   A B C
#> 1 0 0 0
#> 2 0 0 0
#> 
starDesign(k = 3, cc = 2, cs = 2, alpha = "rotatable")
#> $star
#>           A         B         C
#> 1 -1.681793  0.000000  0.000000
#> 2  1.681793  0.000000  0.000000
#> 3  0.000000 -1.681793  0.000000
#> 4  0.000000  1.681793  0.000000
#> 5  0.000000  0.000000 -1.681793
#> 6  0.000000  0.000000  1.681793
#> 
#> $centerStar
#>   A B C
#> 1 0 0 0
#> 2 0 0 0
#> 
#> $centerCube
#>   A B C
#> 1 0 0 0
#> 2 0 0 0
#> 
starDesign(k = 3, cc = 2, cs = 2, alpha = "both")
#> $star
#>           A         B         C
#> 1 -1.732051  0.000000  0.000000
#> 2  1.732051  0.000000  0.000000
#> 3  0.000000 -1.732051  0.000000
#> 4  0.000000  1.732051  0.000000
#> 5  0.000000  0.000000 -1.732051
#> 6  0.000000  0.000000  1.732051
#> 
#> $centerStar
#> NULL
#> 
#> $centerCube
#> NULL
#>