taguchiDesign: Taguchi Designs

Function to create a taguchi design.

taguchiDesign(design, randomize = TRUE, replicates = 1)

Arguments

design

A character string specifying the orthogonal array of the Taguchi design. The available options are:

"L4_2" for three two-level factors.

"L8_2" for seven two-level factors.

"L9_3" for four three-level factors.

"L12_2" for 11 two-level factors.

"L16_2" for 16 two-level factors

"L16_4" for 16 four-level factors.

"L18_2_3" for one two-level and seven three-level factors.

"L25_5" for six five-level factors.

"L27_3" for 13 three-level factors.

"L32_2" for 32 two-level factors.

"L32_2_4" for one two-level factor and nine four-level factors.

"L36_2_3_a" for 11 two-level factors and 12 three-level factors.

"L36_2_3_b" for three two-level factors and 13 three-level factors.

"L50_2_5" for one two-level factor and eleven five-level factors.

"L8_4_2" for one four-level factor and four two-level factors.

"L16_4_2_a" for one four-level factor and 12 two-level factors.

"L16_4_2_b" for two four-level factors and nine two-level factors.

"L16_4_2_c" for three four-level factors and six two-level factors.

"L16_4_2_d" for five four-level factors and two two-level factors.

"L18_6_3" for one six-level factor and six three-level factors.

randomize

A logical value (TRUE/FALSE) that specifies whether to randomize the RunOrder of the design. By default, randomize is set to TRUE.

replicates

An integer specifying the number of replicates for each run in the design.

Value

A taguchiDesign returns an object of class taguchiDesign.

Details

An overview of possible taguchi designs is possible with taguchiChoose.

See also

Examples

tdo <- taguchiDesign("L9_3")
tdo$values(list(A = c("material 1", "material 2", "material 3"), B = c(29, 30, 35)))
tdo$names(c("Factor 1", "Factor 2", "Factor 3", "Factor 4"))
tdo$.response(rnorm(9))
tdo$summary()
#> Taguchi SINGLE Design
#> Information about the factors:
#> 
#>                  A        B        C        D
#> value 1 material 1       29        1        1
#> value 2 material 2       30        2        2
#> value 3 material 3       35        3        3
#> name      Factor 1 Factor 2 Factor 3 Factor 4
#> unit                                         
#> type       numeric  numeric  numeric  numeric
#> 
#> -----------
#> 
#>   StandOrder RunOrder Replicate A B C D   rnorm(9)
#> 1          2        1         1 1 2 2 2 -1.2519859
#> 2          3        2         1 1 3 3 3 -0.5238281
#> 3          7        3         1 3 1 3 2 -0.4968500
#> 4          8        4         1 3 2 1 3 -1.8060313
#> 5          5        5         1 2 2 3 1 -0.5820759
#> 6          1        6         1 1 1 1 1 -1.1088896
#> 7          9        7         1 3 3 2 1 -1.0149620
#> 8          6        8         1 2 3 1 2 -0.1623095
#> 9          4        9         1 2 1 2 3  0.5630558
#> 
#> -----------
#>