taguchiDesign

An R6 class representing a Taguchi experimental design.

Public fields

name

A character string specifying the name of the design. Default is NULL.

factors

A list of factors included in the Taguchi design. Each factor is typically an instance of the taguchiFactor class.

design

A `data.frame` representing the design matrix of the experiment. This includes the levels of each factor for every run of the experiment. Default is an empty data.frame.

designType

A character string specifying the type of Taguchi design used. Default is NULL.

replic

A `data.frame` containing the replication information for the design. Default is an empty data.frame.

response

A `data.frame` storing the response values collected from the experiment. Default is an empty data.frame.

Type

A `data.frame` specifying the type of responses or factors involved in the design. Default is an empty data.frame.

block

A `data.frame` indicating any blocking factors used in the design. Default is an empty data.frame.

runOrder

A `data.frame` detailing the order in which the experimental runs were conducted. Default is an empty data.frame.

standardOrder

A `data.frame` detailing the standard order of the experimental runs. Default is an empty data.frame.

desireVal

A list storing desired values for responses in the experiment. Default is an empty list.

desirability

A list storing desirability functions used to evaluate the outcomes of the experiment. Default is an empty list.

fits

A `data.frame` containing model fits or other statistical summaries from the analysis of the experimental data. Default is an empty data.frame.

Methods

Public methods

taguchiDesign.c$values()

taguchiDesign.c$units()

taguchiDesign.c$.factors()

taguchiDesign.c$names()

taguchiDesign.c$as.data.frame()

taguchiDesign.c$print()

taguchiDesign.c$.response()

taguchiDesign.c$.nfp()

taguchiDesign.c$summary()

taguchiDesign.c$effectPlot()

taguchiDesign.c$identity()

taguchiDesign.c$clone()

Method values()

Get and set the values for an object of class taguchiDesign.

Usage

taguchiDesign.c$values(value)

Arguments

value

New value, If missing value get the values.

Method units()

Get and set the units for an object of class taguchiDesign.

Usage

taguchiDesign.c$units(value)

Arguments

value

New units, If missing value get the units.

Method .factors()

Get and set the factors in an object of class taguchiDesign.

Usage

taguchiDesign.c$.factors(value)

Arguments

value

New factors, If missing value get the factors.

Method names()

Get and set the names in an object of class taguchiDesign.

Usage

taguchiDesign.c$names(value)

Arguments

value

New names, If missing value get the names.

Method as.data.frame()

Return a data frame with the information of the object taguchiDesign.c.

Usage

taguchiDesign.c$as.data.frame()

Method print()

Methods for function print in Package base.

Usage

taguchiDesign.c$print()

Method .response()

Get and set the the response in an object of class taguchiDesign.

Usage

taguchiDesign.c$.response(value)

Arguments

value

New response, If missing value get the response.

Method .nfp()

Prints a summary of the factors attributes including their low, high, name, unit, and type.

Usage

taguchiDesign.c$.nfp()

Method summary()

Methods for function summary in Package base.

Usage

taguchiDesign.c$summary()

Method effectPlot()

Plots the effects of factors on the response variables.

Usage

taguchiDesign.c$effectPlot(
  factors,
  fun = mean,
  response = NULL,
  points = FALSE,
  l.col,
  p.col,
  ld.col,
  lty,
  xlab,
  ylab,
  main,
  ylim,
  pch
)

Arguments

factors

Factors to be plotted.

fun

Function applied to the response variables (e.g., mean).

response

Optional; specifies which response variables to plot.

points

Logical; if TRUE, plots data points.

l.col

A color for the lines.

p.col

A color for the points.

ld.col

A color for the dashed line.

lty

Line type for plotting.

xlab

Label for the x-axis.

ylab

Label for the y-axis.

main

Main title for the plot.

ylim

Limits for the y-axis.

pch

The symbol for plotting points.

Examples

tdo = taguchiDesign("L9_3")
tdo$.response(rnorm(9))
tdo$effectPlot(points = TRUE, col = 2, pch = 16, lty = 3)

Method identity()

Calculates the alias table for a fractional factorial design and prints an easy to read summary of the defining relations such as 'I = ABCD' for a standard 2^(4-1) factorial design.

Usage

taguchiDesign.c$identity()

Method clone()

The objects of this class are cloneable with this method.

Usage

taguchiDesign.c$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `taguchiDesign.c$effectPlot`
## ------------------------------------------------

tdo = taguchiDesign("L9_3")
tdo$.response(rnorm(9))
tdo$effectPlot(points = TRUE, col = 2, pch = 16, lty = 3)
#> Error in tdo$effectPlot(points = TRUE, col = 2, pch = 16, lty = 3): unused argument (col = 2)