adSim: Bootstrap-based Anderson-Darling Test for Univariate

Performs the Anderson-Darling test for univariate distributions with an option for bootstrap-based p-value determination. It also allows p-value determination using tabled critical values.

Usage

adSim(x, distribution = "normal", b = 10000)

Arguments

x

A numeric vector.

distribution

A character string specifying the distribution to test. Recognized distributions include `cauchy`, `exponential`, `gumbel`, `gamma`, `log-normal`, `lognormal`, `logistic`, `normal`, and `weibull`.

b

A numeric value indicating the number of bootstraps to perform. Allowed values range from 1000 to 1,000,000. If b is set to NA, the Anderson-Darling test will be applied without simulation. Note that higher values of b can significantly increase computation time, potentially taking hours depending on the distribution, sample size, and computer system.

Value

A list containing the following components:

distribution

The distribution for which the Anderson-Darling test was applied.

parameter_estimation

The estimated parameters for the distribution.

Anderson_Darling

The value of the Anderson-Darling test statistic.

p_value

The corresponding p-value, either simulated or from tabled values.

critical_values

The critical values corresponding to the 0.75, 0.90, 0.95, 0.975, and 0.99 quantiles, either simulated or from tables.

simAD

The bootstrap-based Anderson-Darling distribution.

Details

The function first estimates the parameters for the tested distribution, typically using Maximum-Likelihood Estimation (MLE) via the FitDistr function. For normal and log-normal distributions, parameters are estimated by the mean and standard deviation. Cauchy distribution parameters are fitted by the sums of the weighted order statistic when using tabled critical values. The Anderson-Darling statistic is then calculated based on these estimated parameters.

Parametric bootstrapping generates the distribution of the Anderson-Darling test statistic, which is used to determine the p-value. This simulation-based Anderson-Darling distribution and its corresponding critical values for selected quantiles can be printed. If simulation is not performed, a p-value is obtained from tabled critical values, although no exact expressions exist except for the log-normal, normal, and exponential distributions.

Examples

x <- rnorm(25, 32, 2)
adSim(x)
#> 
#>     ... simulating the Anderson-Darling distribution by 10000 bootstraps for normal distribution... 
#>  
#> $distribution
#> [1] "normal"
#> 
#> $parameter_estimation
#>      mean        sd 
#> 32.518599  2.093687 
#> 
#> $Anderson_Darling
#> [1] 0.3806083
#> 
#> $p_value
#> [1] 0.3763
#> 
adSim(x, "logistic", 2000)
#> 
#>     ... simulating the Anderson-Darling distribution by 2000 bootstraps for logistic distribution... 
#>  
#> $distribution
#> [1] "logistic"
#> 
#> $parameter_estimation
#> $parameter_estimation$estimate
#>  location     scale 
#> 32.570313  1.242767 
#> 
#> $parameter_estimation$sd
#>  location     scale 
#> 0.4443204 0.2007438 
#> 
#> 
#> $Anderson_Darling
#> [1] 0.4358355
#> 
#> $p_value
#> [1] 0.236
#> 
adSim(x, "cauchy", NA)
#> $distribution
#> [1] "cauchy"
#> 
#> $parameter_estimation
#>  location     scale 
#> 34.028148  1.831827 
#> 
#> $Anderson_Darling
#> [1] 3.917324
#> 
#> $p_value
#> [1] "<= 0.05"
#> 
adSim(x, "exponential", 2000)
#> 
#>     ... simulating the Anderson-Darling distribution by 2000 bootstraps for exponential distribution... 
#>  
#> $distribution
#> [1] "exponential"
#> 
#> $parameter_estimation
#> $parameter_estimation$estimate
#>       rate 
#> 0.03075163 
#> 
#> $parameter_estimation$sd
#>        rate 
#> 0.006150327 
#> 
#> 
#> $Anderson_Darling
#> [1] 10.07883
#> 
#> $p_value
#> [1] 0
#> 
adSim(x, "gumbel", 2000)
#> 
#>     ... simulating the Anderson-Darling distribution by 2000 bootstraps for gumbel distribution... 
#>  
#> $distribution
#> [1] "gumbel"
#> 
#> $parameter_estimation
#> location    scale 
#> 31.48400  1.92032 
#> 
#> $Anderson_Darling
#> [1] 0.587212
#> 
#> $p_value
#> [1] 0.1315
#> 
#x <- rgamma(n = 25, shape = 1)
#adSim(x, "gamma", 2000)
#adSim(x, "log-normal", 2000)
#adSim(x, "lognormal", 2000)
#adSim(x, "weibull", 2000)