gageRRDesign: Gage R&R - Gage Repeatability and Reproducibility

Function to Creates a Gage R&R design.

Usage

gageRRDesign(
  Operators = 3,
  Parts = 10,
  Measurements = 3,
  method = "crossed",
  sigma = 6,
  randomize = TRUE
)

Arguments

Operators

Numeric value giving a number or a character vector defining the Operators. By default Operators is set to `3`.

Parts

A number or character vector defining the Parts. By default parts is set to `10`.

Measurements

A number defining the measurements per part. By default Measurements is set to `3`.

method

Character string specifying the Gage R&R method. `crossed` which is the typical design for performing a Measurement Systems Analysis using Gage Repeatability and Reproducibility or `nested` which is used for destructive testing (i.e. the same part cannot be measured twice). Operators measure each a different sample of parts under the premise that the parts of each batch are alike. By default method is set to `crossed`.

sigma

For sigma=6 this relates to 99.73 percent representing the full spread of a normal distribution function (i.e. pnorm(3) - pnorm(-3)). Another popular setting sigma=5.15 relates to 99 percent (i.e. pnorm(2.575) - pnorm(-2.575)). By default sigma is set to `6`.

randomize

Logical value. TRUE (default) randomizes the gageRR design.

Value

The function gageRRDesign returns an object of class gageRR.

See also

Examples

design <- gageRRDesign(Operators = 3, Parts = 10, Measurements = 3,
                       method = "crossed", sigma = 6, randomize = TRUE)