mixDesign-class: Class "mixDesign"

mixDesign class for simplex lattice and simplex centroid mixture designs with optional center points and augmented points.

See also

Public fields

name

Character string representing the name of the design.

factors

List of factors involved in the mixture design, including their levels and settings.

total

Numeric value representing the total number of runs in the design.

lower

Numeric vector representing the lower bounds of the factors in the design.

design

Data frame containing the design matrix for the mixture design.

designType

Character string specifying the type of design (e.g., "simplex-lattice", "simplex-centroid").

pseudo

Data frame containing pseudo-experimental runs if applicable.

response

Data frame containing the responses or outcomes measured in the design.

Type

Data frame specifying the type of design used (e.g., "factorial", "response surface").

block

Data frame specifying block structures if the design is blocked.

runOrder

Data frame specifying the order in which runs are performed.

standardOrder

Data frame specifying the standard order of the runs.

desireVal

List of desired values or targets for the response variables.

desirability

List of desirability scores or metrics based on the desired values.

fits

Data frame containing the fitted model parameters and diagnostics.

Methods

Public methods

mixDesign.c$.factors()

mixDesign.c$names()

mixDesign.c$as.data.frame()

mixDesign.c$print()

mixDesign.c$.response()

mixDesign.c$.nfp()

mixDesign.c$summary()

mixDesign.c$units()

mixDesign.c$lows()

mixDesign.c$highs()

mixDesign.c$clone()

Method .factors()

Get and set the factors in an object of class mixDesign

Usage

mixDesign.c$.factors(value)

Arguments

value

New factors, If missing value get the factors.

Method names()

Get and set the names in an object of class mixDesign.

Usage

mixDesign.c$names(value)

Arguments

value

New names, If missing value get the names.

Method as.data.frame()

Methods for function as.data.frame in Package base.

Usage

mixDesign.c$as.data.frame()

Method print()

Methods for function print in Package base.

Usage

mixDesign.c$print()

Method .response()

Get and set the the response in an object of class mixDesign.

Usage

mixDesign.c$.response(value)

Arguments

value

New response, If missing value get the response.

Method .nfp()

Prints a summary of the factors attributes including their low, high, name, unit, and type.

Usage

mixDesign.c$.nfp()

Method summary()

Methods for function summary in Package base.

Usage

mixDesign.c$summary()

Method units()

Get and set the units for the factors in an object of class mixDesign.

Usage

mixDesign.c$units(value)

Arguments

value

New units, If missing value get the units.

Method lows()

Get and set the lows for the factors in an object of class mixDesign.

Usage

mixDesign.c$lows(value)

Arguments

value

New lows, If missing value get the lows.

Method highs()

Get and set the highs for the factors in an object of class mixDesign.

Usage

mixDesign.c$highs(value)

Arguments

value

New highs, If missing value get the highs.

Method clone()

The objects of this class are cloneable with this method.

Usage

mixDesign.c$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.