steepAscent: Steepest Ascent

steepAscent is a method to calculate the steepest ascent for a facDesign.c object.

Usage

steepAscent(factors, response, size = 0.2, steps = 5, data)

Arguments

factors

List containing vector of factor names (coded) to be included in calculation, first factor is the reference factor.

response

A character of response given in data.

size

Numeric integer value giving the step size in coded units for the first factor given in factors. By default size is set to 0.2.

steps

Numeric integer value giving the number of steps. By default step is set to `5`.

data

An object of class facDesign.c.

Value

steepAscent returns an object of class steepAscent.c.

See also

Examples

# Example 1
fdo = facDesign(k = 2, centerCube = 5)
fdo$lows(c(170, 150))
fdo$highs(c(230, 250))
fdo$names(c("temperature", "time"))
fdo$unit(c("C", "minutes"))
yield = c(32.79, 24.07, 48.94, 52.49, 38.89, 48.29, 29.68, 46.5, 44.15)
fdo$.response(yield)
fdo$summary()
#> Information about the factors:
#> 
#>      temperature    time
#> low          170     150
#> high         230     250
#> name temperature    time
#> unit           C minutes
#> type     numeric numeric
#> -----------
#>   StandOrder RunOrder Block  A  B yield
#> 4          4        1     1  1  1 32.79
#> 6          6        2     1  0  0 24.07
#> 9          9        3     1  0  0 48.94
#> 3          3        4     1 -1  1 52.49
#> 2          2        5     1  1 -1 38.89
#> 1          1        6     1 -1 -1 48.29
#> 7          7        7     1  0  0 29.68
#> 5          5        8     1  0  0 46.50
#> 8          8        9     1  0  0 44.15

sao = steepAscent(factors = c("B", "A"), response = "yield", size = 1,
                  data = fdo)
#> 
#> Steepest Ascent for fdo 
#> 
#>   Run Delta B.coded A.coded B.real A.real
#> 1   1     0       0     0.0    200    200
#> 2   2     1      -1   -15.3    150   -259
#> 3   3     2      -2   -30.6    100   -719
#> 4   4     3      -3   -45.9     50  -1178
#> 5   5     4      -4   -61.3      0  -1638
#> 6   6     5      -5   -76.6    -50  -2097