steepAscent-class: Class "steepAscent"

The steepAscent.c class represents a steepest ascent algorithm in a factorial design context. This class is used for optimizing designs based on iterative improvements.

See also

Public fields

name

A character string representing the name of the steep ascent design.

X

A data frame containing the design matrix for the steepest ascent procedure. This matrix represents the factors and their levels at each iteration.

response

A data frame containing the response values associated with the design matrix.

Methods

Public methods

steepAscent.c$.response()

steepAscent.c$get()

steepAscent.c$as.data.frame()

steepAscent.c$print()

steepAscent.c$plot()

steepAscent.c$clone()

Method .response()

Get and set the `response` values in an object of class `steepAscent.c`.

Usage

steepAscent.c$.response(value)

Arguments

value

A data frame or numeric vector to set as the new `response`. If missing, returns the current `response`.

Method get()

Access specific elements in the design matrix or response data of the object.

Usage

steepAscent.c$get(i, j)

Arguments

i

An integer specifying the row index to retrieve.

j

An integer specifying the column index to retrieve.

Method as.data.frame()

Convert the object to a data frame.

Usage

steepAscent.c$as.data.frame()

Method print()

Print the details of the object.

Usage

steepAscent.c$print()

Method plot()

Plot the results of the steepest ascent procedure for an object of class `steepAscent.c`.

Usage

steepAscent.c$plot(main, xlab, ylab, l.col, p.col, line.type, point.shape)

Arguments

main

The main title of the plot.

xlab

The label for the x-axis.

ylab

The label for the y-axis.

l.col

Color for the line in the plot.

p.col

Color for the points in the plot.

line.type

Type of the line used in the plot.

point.shape

Shape of the points used in the plot.

Method clone()

The objects of this class are cloneable with this method.

Usage

steepAscent.c$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.